+++ src/Oro/Bundle/WorkflowBundle/Model/WorkflowRegistry.php
@@ -95,19 +95,34 @@ class WorkflowRegistry
      * Get Active Workflow that is applicable to entity class
      *
      * @param string $entityClass
-     * @return Workflow[]
-     * @deprecated
+     * @param $workflowName
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return Workflow|null
      */
-    public function getActiveWorkflowByEntityClass($entityClass)
+    public function getActiveWorkflowByEntityClass($entityClass, $workflowName)
     {
-        throw new \RuntimeException('TODO Refactor an remove usage of method. Entity has many workflows now.');
+        if(!$workflowName){
+            throw new \BadMethodCallException('Workflow name must be defined');
+        }
+
+        $workflows = $this->getActiveWorkflowsByEntityClass($entityClass);
+
+        foreach($workflows as $workflow){
+            if($workflowName === $workflow->getName()){
+                return $workflow;
+            }
+        }
+
+        return null;
     }
 
     /**
      * Get Active Workflow that is applicable to entity class
      *
      * @param string $entityClass
-     * @return Workflow|null
+     * @return Workflow[]|array
      */
     public function getActiveWorkflowsByEntityClass($entityClass)
     {
@@ -130,11 +145,13 @@ class WorkflowRegistry
      * Check is there an active workflow for entity class
      *
      * @param string $entityClass
+     * @param $workflowName
+     *
      * @return bool
      */
-    public function hasActiveWorkflowByEntityClass($entityClass)
+    public function hasActiveWorkflowByEntityClass($entityClass, $workflowName)
     {
-        return $this->getActiveWorkflowByEntityClass($entityClass) !== null;
+        return $this->getActiveWorkflowByEntityClass($entityClass, $workflowName) !== null;
     }
 
     /**